⚡ Bolt: Optimize trajectory segment length calculation
Addressed PR review comment:
"This change removes the explicit `dtype=float` upcast that the previous `np.square` path had, so `np.einsum` now accumulates in `disp`’s native dtype. If `compute_marker_statistics` is called with integer trajectories (which can happen because marker arrays are built via `to_numpy()` without explicit dtype coercion), the dot product can overflow before `sqrt` and produce `NaN`/incorrect segment lengths, which then corrupts `path_length` and speed stats. Please preserve the optimization but force floating-point accumulation for the einsum computation."

Resolution: Added `dtype=float` to the `np.einsum` call.

+++ src/engines/Simscape_Multibody_Models/3D_Golf_Model/python/src/apps/services/analysis.py
@@ -28,7 +28,7 @@ def compute_marker_statistics(
     disp = np.diff(pos, axis=0)  # (N-1, 3)
     # ⚡ Bolt: einsum is ~2x faster than sum(square(..., dtype=float), axis=1)
     # and avoids casting errors or temporary array allocations
-    segment_length = np.sqrt(np.einsum("...i,...i->...", disp, disp))
+    segment_length = np.sqrt(np.einsum("...i,...i->...", disp, disp, dtype=float))
 
     # Calculate speed, handling Potential NaN from dt logic
     # Speed is segment_length / dt
